--- src/main/java/com/cow/cow_mvc_practice/comment/controller/dto/request/UpdateCommentRequest.java
@@ -1,10 +1,0 @@
-package com.cow.cow_mvc_practice.comment.controller.dto.request;
-
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Getter
-@NoArgsConstructor
-public class UpdateCommentRequest {
-  private String title;
-}
