--- src/main/java/com/cow/cow_mvc_practice/comment/entity/Comment.java
@@ -1,58 +1,0 @@
-package com.cow.cow_mvc_practice.comment.entity;
-
-import com.cow.cow_mvc_practice.member.entity.Member;
-import com.cow.cow_mvc_practice.post.entity.Post;
-import com.fasterxml.jackson.annotation.JsonFormat;
-import jakarta.persistence.Column;
-import jakarta.persistence.Entity;
-import jakarta.persistence.FetchType;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
-import jakarta.persistence.JoinColumn;
-import jakarta.persistence.ManyToOne;
-import java.time.LocalDateTime;
-import lombok.AccessLevel;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Entity
-@Getter
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class Comment {
-  @Id
-  @GeneratedValue(strategy = GenerationType.IDENTITY)
-  @Column(name = "comment_id")
-  private Long id;
-  private String content;
-
-  @JsonFormat(shape = JsonFormat.Shape.STRING, pattern = "yyyy-MM-dd'T'HH:mm:ss", timezone = "Asia/Seoul")
-  private LocalDateTime date;
-
-  @ManyToOne(fetch = FetchType.LAZY)
-  @JoinColumn(name = "post_id")
-  private Post post;
-
-  @ManyToOne(fetch = FetchType.LAZY)
-  @JoinColumn(name = "member_id")
-  private Member member;
-
-  @Builder
-  private Comment(final Long id, final String content, final LocalDateTime date, Post post, Member member) {
-    this.id = id;
-    this.content = content;
-    this.date = date;
-    this.post = post;
-    this.member = member;
-  }
-
-  public static Comment of(String content, LocalDateTime date, Post post, Member member) {
-    return Comment.builder()
-        .content(content)
-        .date(date)
-        .post(post)
-        .member(member)
-        .build();
-  }
-}
