--- src/main/java/com/cow/cow_mvc_practice/member/controller/dto/request/UpdateMemberRequest.java
@@ -1,10 +1,0 @@
-package com.cow.cow_mvc_practice.member.controller.dto.request;
-
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Getter
-@NoArgsConstructor
-public class UpdateMemberRequest {
-	private String name;
-}
