--- src/main/java/com/cow/cow_mvc_practice/member/repository/MemberRepository.java
@@ -1,13 +1,0 @@
-package com.cow.cow_mvc_practice.member.repository;
-
-import java.util.List;
-import java.util.Optional;
-
-import com.cow.cow_mvc_practice.member.entity.Member;
-
-public interface MemberRepository {
-
-	void save(Member member);
-	Optional<Member> findById(Long memberId);
-	Optional<List<Member>> findAll();
-}
