--- src/main/java/com/cow/cow_mvc_practice/post/repository/PostRepositoryImpl.java
@@ -1,27 +1,0 @@
-package com.cow.cow_mvc_practice.post.repository;
-
-import com.cow.cow_mvc_practice.post.entity.Post;
-import java.util.HashMap;
-import java.util.List;
-import java.util.Map;
-import java.util.Optional;
-
-public class PostRepositoryImpl implements PostRepository{
-
-  private static final Map<Long, Post> store = new HashMap<>();
-
-  @Override
-  public void save(Post post) {
-    store.put(post.getId(), post);
-  }
-
-  @Override
-  public Optional<Post> findById(Long memberId) {
-    return Optional.ofNullable(store.get(memberId));
-  }
-
-  @Override
-  public Optional<List<Post>> findAll() {
-    return Optional.of((List<Post>)store.values());
-  }
-}
