+++ src/main/java/com/cow/mvc/comment/controller/dto/request/UpdateCommentRequest.java
@@ -1,0 +1,10 @@
+package com.cow.mvc.comment.controller.dto.request;
+
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Getter
+@NoArgsConstructor
+public class UpdateCommentRequest {
+	private String title;
+}
