+++ src/main/java/com/cow/mvc/member/controller/dto/request/MemberRequest.java
@@ -1,0 +1,18 @@
+package com.cow.mvc.member.controller.dto.request;
+
+import com.cow.mvc.member.entity.Member;
+
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Getter
+@NoArgsConstructor
+public class MemberRequest {
+	private String name;
+
+	public Member toEntity() {
+		return Member.builder()
+			.name(this.name)
+			.build();
+	}
+}
