+++ src/main/java/com/cow/mvc/member/controller/dto/request/UpdateMemberRequest.java
@@ -1,0 +1,10 @@
+package com.cow.mvc.member.controller.dto.request;
+
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Getter
+@NoArgsConstructor
+public class UpdateMemberRequest {
+	private String name;
+}
