+++ src/main/java/com/cow/mvc/member/repository/MemberJPARepository.java
@@ -1,0 +1,8 @@
+package com.cow.mvc.member.repository;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import com.cow.mvc.member.entity.Member;
+
+public interface MemberJPARepository extends JpaRepository<Member, Long> {
+}
