+++ src/main/java/com/cow/mvc/post/controller/dto/request/PostRequest.java
@@ -1,0 +1,11 @@
+package com.cow.mvc.post.controller.dto.request;
+
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Getter
+@NoArgsConstructor
+public class PostRequest {
+	private String title;
+	private String content;
+}
