+++ src/main/java/com/cow/mvc/post/entity/Post.java
@@ -1,0 +1,63 @@
+package com.cow.mvc.post.entity;
+
+import java.time.LocalDateTime;
+import java.util.ArrayList;
+import java.util.List;
+
+import com.cow.mvc.comment.entity.Comment;
+import com.cow.mvc.member.entity.Member;
+import com.fasterxml.jackson.annotation.JsonFormat;
+
+import jakarta.persistence.CascadeType;
+import jakarta.persistence.Column;
+import jakarta.persistence.Entity;
+import jakarta.persistence.FetchType;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
+import jakarta.persistence.JoinColumn;
+import jakarta.persistence.ManyToOne;
+import jakarta.persistence.OneToMany;
+import lombok.AccessLevel;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Getter
+@NoArgsConstructor(access = AccessLevel.PROTECTED) // 생성자를 통해서 값 변경 목적으로 접근하는 메시지들 차단
+@Entity
+public class Post {
+	@Id
+	@Column(name = "post_id")
+	@GeneratedValue(strategy = GenerationType.IDENTITY)
+	private Long id;
+	private String title;
+	private String content;
+
+	@JsonFormat(shape = JsonFormat.Shape.STRING, pattern = "yyyy-MM-dd'T'HH:mm:ss", timezone = "Asia/Seoul")
+	private LocalDateTime date;
+
+	@ManyToOne(fetch = FetchType.LAZY, cascade = CascadeType.ALL)
+	@JoinColumn(name = "member_id")
+	private Member member;
+
+	@OneToMany(mappedBy = "post", cascade = CascadeType.ALL)
+	private final List<Comment> comments = new ArrayList<>();
+
+	@Builder
+	private Post(final String title, final String content, final Member member, final LocalDateTime date) {
+		this.title = title;
+		this.content = content;
+		this.member = member;
+		this.date = date;
+	}
+
+	public static Post of(final String title, final String content, final Member member, final LocalDateTime date) {
+		return Post.builder()
+			.title(title)
+			.content(content)
+			.member(member)
+			.date(date)
+			.build();
+	}
+}
